Revert "refactor: move methods from dto to energy function"
This reverts commit 4547d6f78ff98bca394820b987f8b94680b20e5d.
