+++ src/libecalc/domain/infrastructure/energy_components/legacy_consumer/tabulated/tabular_energy_function.py
@@ -1,53 +1,33 @@
 from __future__ import annotations
 
-from typing import Literal
-
 import numpy as np
 from scipy.interpolate import LinearNDInterpolator, interp1d
 
-from libecalc.common.energy_model_type import EnergyModelType
 from libecalc.common.energy_usage_type import EnergyUsageType
-from libecalc.common.errors.exceptions import IllegalStateException, InvalidColumnException
+from libecalc.common.errors.exceptions import IllegalStateException
 from libecalc.common.list.adjustment import transform_linear
 from libecalc.common.list.list_utils import array_to_list
 from libecalc.common.logger import logger
 from libecalc.common.units import Unit
-from libecalc.domain.component_validation_error import (
-    ModelValidationError,
-    ProcessEqualLengthValidationException,
-    ProcessHeaderValidationException,
-)
 from libecalc.domain.infrastructure.energy_components.legacy_consumer.tabulated.common import Variable
 from libecalc.domain.process.core.results.base import EnergyFunctionResult
-from libecalc.presentation.yaml.validation_errors import Location
+from libecalc.domain.process.dto import TabulatedData
 
 
 class TabularEnergyFunction:
-    typ: Literal[EnergyModelType.TABULATED] = EnergyModelType.TABULATED
-
     def __init__(
         self,
-        headers: list[str],
-        data: list[list[float]],
-        energy_usage_adjustment_constant: float,
-        energy_usage_adjustment_factor: float,
+        energy_model: TabulatedData,
     ):
-        self.headers = headers
-        self.data = data
-        self.energy_usage_adjustment_constant = energy_usage_adjustment_constant
-        self.energy_usage_adjustment_factor = energy_usage_adjustment_factor
-        self.validate_headers()
-        self.validate_data()
-
         """Tabular consumer energy function [MW] or [Sm3/day]."""
-        function_values = self.get_function_values()
-        variables = [Variable(name=name, values=values) for name, values in self.get_variables().items()]
+        function_values = energy_model.get_function_values()
+        variables = [Variable(name=name, values=values) for name, values in energy_model.get_variables().items()]
 
         self.required_variables = [variable.name for variable in variables]
         function_values_adjusted = transform_linear(
             values=np.reshape(np.asarray(function_values), -1),
-            constant=self.energy_usage_adjustment_constant,
-            factor=self.energy_usage_adjustment_factor,
+            constant=energy_model.energy_usage_adjustment_constant,
+            factor=energy_model.energy_usage_adjustment_factor,
         )
         if len(variables) == 1:
             self._func = interp1d(
@@ -63,64 +43,7 @@ class TabularEnergyFunction:
                 fill_value=np.nan,
                 rescale=True,
             )
-        self.energy_usage_type = self.get_energy_usage_type()
-
-    def interpolate(self, variables_array: np.ndarray) -> np.ndarray:
-        """
-        Public method to interpolate energy usage.
-        Args:
-            variables_array: np.ndarray of variable values, shape (n_points, n_variables) or (n_variables,)
-        Returns:
-            np.ndarray: Interpolated energy usage values.
-        """
-        return self._func(variables_array)
-
-    def get_column(self, header: str) -> list:
-        return self.data[self.headers.index(header)]
-
-    def get_energy_usage_type(self) -> EnergyUsageType:
-        return EnergyUsageType.POWER if EnergyUsageType.POWER.value in self.headers else EnergyUsageType.FUEL
-
-    def _get_function_value_header(self) -> str:
-        return self.get_energy_usage_type().value
-
-    def get_function_values(self) -> list[float]:
-        return self.get_column(self._get_function_value_header())
-
-    def get_variables(self) -> {str, list[float]}:
-        variable_headers = [header for header in self.headers if header != self._get_function_value_header()]
-        return {header: self.get_column(header) for header in variable_headers}
-
-    def validate_headers(self):
-        is_valid_headers = len(self.headers) > 0 and ("FUEL" in self.headers or "POWER" in self.headers)
-        if not is_valid_headers:
-            msg = "TABULAR facility input type data must have a 'FUEL' or 'POWER' header"
-
-            raise ProcessHeaderValidationException(
-                errors=[
-                    ModelValidationError(name=self.typ.value, location=Location([self.typ.value]), message=str(msg))
-                ],
-            )
-
-    def validate_data(self):
-        lengths = [len(lst) for lst in self.data]
-        if len(set(lengths)) > 1:
-            problematic_vectors = [(i, len(lst)) for i, lst in enumerate(self.data)]
-            msg = f"TABULAR facility input type data should have equal number of datapoints for all headers. Found lengths: {problematic_vectors}"
-
-            raise ProcessEqualLengthValidationException(
-                errors=[
-                    ModelValidationError(name=self.typ.value, location=Location([self.typ.value]), message=str(msg))
-                ],
-            )
-        for column_index, header in enumerate(self.headers):
-            for row_index, value in enumerate(self.data[column_index]):
-                try:
-                    float(value)
-                except ValueError as e:
-                    raise InvalidColumnException(
-                        header=header, message=f"Got non-numeric value '{value}'.", row=row_index
-                    ) from e
+        self.energy_usage_type = energy_model.get_energy_usage_type()
 
     def evaluate_variables(self, variables: list[Variable]) -> EnergyFunctionResult:
         variables_map_by_name = {variable.name: variable.values for variable in variables}
